requirements.txt
deprecation==2.1.0
et-xmlfile==1.1.0
img2pdf==0.4.4
lxml==4.9.2
numpy==1.25.0
opencv-python==4.5.5.62
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pathlib==1.0.1
pdf2img==0.1.2
pikepdf==7.2.0
Pillow==9.5.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
scipy==1.10.1
six==1.16.0
tzdata==2023.3
watchdog==3.0.0
Unidecode==1.3.6
python-dotenv==1.0.0